+++ hephaistos/interactive.py
@@ -7,21 +7,28 @@ CANCEL_OPTION='Cancel'
 EXIT_OPTION='Exit'
 
 
-def pick(prompt: str="Pick an option:", options: Union[dict, list]=None, add_option: str=CANCEL_OPTION, *args, **kwargs) -> Any:
-    if not options and kwargs:
-        options = kwargs
-    if add_option:
-        if isinstance(options, dict):
-            options[add_option] = add_option
-        elif add_option not in options:
-            options.append(add_option)
+def pick(prompt: str="Pick an option:", options: Union[dict, list]=None, recommended_option: str=None, additional_option: str=CANCEL_OPTION) -> Any:
+    # transform options list to dict (we allow options to be passed in for comfort but only want to handle dicts in code)
+    if isinstance(options, list):
+        options = {option: option for option in options}
+    # prepend recommended tag to recommended option
+    recommended_tag = "[RECOMMENDED]"
+    if recommended_option and options[recommended_option] and not options[recommended_option].startswith(recommended_tag):
+        options[recommended_option] = f"{recommended_tag} {options[recommended_option]}"
+    # add additional option
+    if additional_option:
+        options[additional_option] = additional_option
     print(prompt)
     index_key_dict = {index + 1: key for index, key in enumerate(options)}
     for index, key in index_key_dict.items():
-        print(f"{index}. {options[key]}") if isinstance(options, dict) else print(f"{index}. {key}")
-    print("Choice: ", end='', flush=True)
+        print(f"{index}. {options[key]}")
+    msg = "Choice: " if not recommended_option else f"Choice (press ENTER to use {recommended_tag}): "
+    print(msg, end='', flush=True)
     char = getch()
     print(char + '\n')
+    # if user pressed enter and a recommended option was set, use it
+    if char == '\r' and recommended_option:
+        return recommended_option
     try:
         option = int(char)
         value = index_key_dict[option]
@@ -31,8 +38,10 @@ def pick(prompt: str="Pick an option:", options: Union[dict, list]=None, add_opt
             raise InteractiveExit()
         return index_key_dict[option]
     except (ValueError, KeyError):
-        print("Please input a valid number from the given list.")
-        return pick(prompt, options, False)
+        msg = "Please input a valid number from the given list"
+        msg += f", or press ENTER to use to use the {recommended_tag} option." if recommended_option else "."
+        print(f"{msg}\n")
+        return pick(prompt=prompt, options=options, recommended_option=recommended_option, additional_option=None)
 
 
 class InteractiveCancel(RuntimeError): ...